+++ R/post-file.R
@@ -1,7 +1,7 @@
 #' Post a file to file.io
 #'
 #' @md
-#' @param path path to a file
+#' @param path path to a file (it will be [path.expand()]ed and [normalizePath()]'d)
 #' @param expires defaults to 14 days (`14d`). Must be a positive integer which,
 #'        by default, represents the number of days until the file will be deleted.
 #'        If you follow it with `w`, it will be the number of weeks; `m` for months;
@@ -12,8 +12,10 @@
 #' @references <https://www.file.io/>
 #' @note There is a 5GB per file limit for the free version.
 #' @examples
-#' fi_post(system.file("extdat", "tst.txt", package = "fileio"))
-fi_post <- function(path, expires="14d") {
+#' x <- fi_post_file(system.file("extdat", "tst.txt", package = "fileio"))
+#' readLines(con <- url(x$link), warn = FALSE)
+#' close(con)
+fi_post_file <- function(path, expires = "14d") {
 
   if (!grepl("[[:digit:]]+[wdmy]", expires[1])) {
     stop("'expires' must be either an integer or an integer followed by one of [dwmy]")
@@ -24,17 +26,20 @@ fi_post <- function(path, expires="14d") {
 
   if (!file.exists(path)) stop("File not found!", call.=FALSE)
 
+  if (file.size(path) > 5000000000) stop("file.io has a 5GB limit on uploads.", call.=FALSE)
+
   httr::POST(
     url = "https://file.io",
     encode = "multipart",
-    body = list(file=httr::upload_file(path)),
+    body = list(file = httr::upload_file(path)),
+    httr::user_agent(FILEIO_USER_AGENT)
   ) -> res
 
   httr::stop_for_status(res)
 
   res <- httr::content(res)
 
-  res <- as.data.frame(res, stringsAsFactors=FALSE)
+  res <- as.data.frame(res, stringsAsFactors = FALSE)
 
   class(res) <- c("tbl_df", "tbl", "data.frame")
 
@@ -42,3 +47,6 @@ fi_post <- function(path, expires="14d") {
 
 }
 
+#' @noRd
+#' @export
+fi_post <- fi_post_file